Drop steps from fedora 38 runner
There are a bunch of failures when these commands run and it's annoying.
I started a thread in the ImageBuilder matrix channel to try to
understand why we can't override the runner for a non 'build' pipeline.
For now let's just delete the calls because AFAICT we don't need them.