osbuild/mounts: adjust the source path to use when mounting devices
Adjust the source path to just have /dev in front (i.e. /dev/loop0)
and not be a path to the temporary dev directory that was prepared.
We do this because some tools (like grub2-install) consult
mountinfo to try to canonicalize paths for mounts.

Fixes https://github.com/osbuild/osbuild/issues/1492