stages/coreos.live-artifacts: extract test fixture handling
Extract functions that perform actions related to the existence of the
text fixtures, which presumably makes the stage behave in a test mode,
not using real image files.

These should eventually be removed and made part of a unit or
integration test for the stage.  A set of non-monolithic stages would
make it possible to swap out parts that set up images for ones that set
up test fixtures and allow us to test atomic functionality separately.